gettingFilmsForQuery
Получаю по кнопке список фильмов и создаю новую разметку . Начальный варик, без условий,без проверок. С АПИ не все картинки подгружаются, некоторые данные сервис не дает
