+++ src/partials/testForm.html
@@ -1,5 +1,5 @@
 //Форма временная
 <form action="" class="search-form" id="search-form">
   <input type="search" class="input-area" />
-  <i class="fa fa-search search-button"></i>
+  <button class="search-button">ПОИСК ФИЛЬМА</button>
 </form>
