+++ src/templates/mainListOfFilms.hbs
@@ -1,0 +1,8 @@
+<li class="card">
+    <a href="">
+            <img src="https://www.themoviedb.org/t/p/w300_and_h450_bestv2{{poster_path}}" alt={{title}} />
+            <p>Название фильма  {{original_title}}</p>
+            <p>АйДи жанров  {{genre_ids}}</p>
+            <p>Дата релиза  {{release_date}}</p>
+    </a>
+</li>
